Added edge_collapse to the triangulated mixed complex

+++ Skin_surface_3/examples/Skin_surface_3/dsrpdb/lib/PDB.cc
@@ -82,7 +82,7 @@ namespace dsrpdb {
     assert(i < models_.size());
     return models_[i];
   };
-  unsigned int PDB::number_of_models() const {
+  size_t PDB::number_of_models() const {
     return models_.size(); 
   }
